advection fixed gpu, debugging cmake, update tests, try to find issue of pressure correciton

+++ include/utils.cuh
@@ -30,7 +30,7 @@ void print_matrix(const int &m, const int &n, const double *A, const int &lda);
 
 void print_matrix_row_major(const int &m, const int &n, const double *A, const int &lda); 
 
-__host__ __device__ int periodic_linear_Idx(const int &x, const int &y, const int bound_x = 2*NUM_N, const int bound_y = M);
+__host__ __device__ size_t periodic_linear_Idx(const int &x, const int &y, const int bound_x = 2*NUM_N, const int bound_y = M);
 
 __host__ __device__ void setClosestGridPointIdx(double x, double y, int n, int m, int &closest_x_i, int &closest_y_i);
